fix: reorder tensor conversion before .float() call in ConstraintManager.compute()
The .float() call was happening before checking if the return value from
term_cfg.func was a Tensor, making the isinstance check and conversion
branch unreachable. Non-Tensor returns (plain Python lists, ints, floats,
numpy arrays) would raise AttributeError instead of being converted.

Fixes the ordering: convert first (if needed), then call .float().
Adds test_constraint_term_converts_plain_python_values_to_tensor to
verify the previously unreachable conversion path now works correctly.

+++ talon_rl/isaaclab/managers/constraint_manager.py
@@ -172,10 +172,12 @@ class ConstraintManager(ManagerBase):
         self._delta_buf.zero_()
 
         for name, term_cfg in zip(self._term_names, self._term_cfgs):
-            value = term_cfg.func(self._env, **term_cfg.params).float()
+            value = term_cfg.func(self._env, **term_cfg.params)
 
             if not isinstance(value, torch.Tensor):
                 value = torch.tensor(value, device=self.device, dtype=torch.float32)
+
+            value = value.float()
 
             if term_cfg.time_out == "truncate":
                 value = torch.clamp(value, 0.0, 1.0)
